+++ libraries/TSL235R/CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.4] - 2023-02-18
+- updated cf value in **calculateFactor()**
+- add **#define TSL235_DEFAULT_VOLTAGE** to allow set voltage from command line.
+- update readme.md
+- move code to .cpp
+- update GitHub actions
+- update license 2023
+- minor edits
+
+
 ## [0.1.3] - 2022-11-26
 - Add RP2040 support to build-CI.
 - Add CHANGELOG.md